Version compiling

+++ kmsr/cpp/gonzales.cpp
@@ -1,8 +1,7 @@
-#include "header/gonzales.h"
-
 #include <algorithm>
 
-#include "header/point.h"
+#include "../header/gonzales.h"
+#include "../header/point.h"
 
 using namespace std;
 
@@ -16,12 +15,12 @@ double gonzalesrmax(const vector<Point> &points, int k) {
   vector<double> distances(points.size());
 
   // Berechne den maximalen Abstand für die initiale Zuordnung
-  for (int i = 0; i < points.size(); i++) {
+  for (size_t i = 0; i < points.size(); i++) {
     distances[i] = Point::distance(points[i], centers[0]);
   }
 
   // Iteriere, um die restlichen k - 1 Zentren zu finden
-  while (centers.size() < k) {
+  while (static_cast<int>(centers.size()) < k) {
     // Wähle das nächste Zentrum basierend auf dem größten Abstand
     int nextCenterIndex = distance(
         distances.begin(), max_element(distances.begin(), distances.end()));
@@ -29,14 +28,14 @@ double gonzalesrmax(const vector<Point> &points, int k) {
     centers.push_back(nextCenter);
 
     // Aktualisiere die Abstände für jeden Punkt zu seinem nächsten Zentrum
-    for (int i = 0; i < points.size(); i++) {
+    for (size_t i = 0; i < points.size(); i++) {
       distances[i] = min(distances[i], Point::distance(points[i], nextCenter));
     }
   }
 
   // Berechne den maximalen Abstand der Punkte zu ihren Zentren
   double maxDistance = 0.0;
-  for (int i = 0; i < points.size(); i++) {
+  for (size_t i = 0; i < points.size(); i++) {
     maxDistance = max(maxDistance, distances[i]);
   }
